改进服务器代码架构，FIX服务器响应登录包的bug
