+++ README.MD
@@ -2,6 +2,6 @@
 
 视频地址：[Golang基础-进阶-实战编程](https://www.bilibili.com/video/BV1kt411C7fK)
 
-学习进度：318 /389 (286-291未学习)
+学习进度：321 /389 (286-291未学习)
 
 持续更新，督促自己学习golang
